Migrate server files and folders to /server, change frontend folder and related scripts to client, move documentation to /documentation
