+++ .gitignore
@@ -2,4 +2,5 @@
 /node_modules
 /config/keys.js
 /frontend/node_modules
+.sass-cache
 /.sass-cache
