+++ client/src/assets/stylesheets/card_anim.scss
