+++ client/src/assets/stylesheets/nav_bar.scss
