+++ client/src/assets/stylesheets/splash.scss
