+++ client/src/components/splash/Splash.jsx
