+++ client/src/reducers/errors_reducer.js
