+++ client/src/reducers/session_errors_reducer.js
