+++ client/src/reducers/session_reducer.js
