+++ client/src/util/session_api_util.js
