+++ server/config/keys.js
@@ -1,0 +1,4 @@
+module.exports = {
+  mongoURI: 'mongodb+srv://StevenKT:nCo4P9nqlStCfJfM@cah.bo7wx.mongodb.net/<dbname>?retryWrites=true&w=majority',
+  secretOrKey: "@*a7+tzV9^':d27roNH[[}I-)LJI.u"
+}
